+++ packages/components/organisms/f-checkout/src/components/_tests/Checkout.test.js
@@ -1255,6 +1255,17 @@ describe('Checkout', () => {
                     });
                 });
 
+                it('should make a call to `toggleDialogError`', () => {
+                    // Arrange
+                    const toggleDialogErrorSpy = jest.spyOn(wrapper.vm, 'toggleDialogError');
+
+                    // Act
+                    wrapper.vm.handleNonFulfillableCheckout();
+
+                    // Assert
+                    expect(toggleDialogErrorSpy).toHaveBeenCalled();
+                });
+
                 it('should make a call to `trackFormErrors`', () => {
                     // Arrange
                     const trackFormErrorsSpy = jest.spyOn(wrapper.vm, 'trackFormErrors');
@@ -2735,6 +2746,38 @@ describe('Checkout', () => {
 
                     result.rejects.toThrow(PlaceOrderError);
                     result.rejects.toThrow(errorMessage);
+                });
+
+                it('should call `toggleDialogError`', async () => {
+                    // Arrange
+                    const errorMessage = 'An error';
+
+                    wrapper = shallowMount(VueCheckout, {
+                        store: createStore(
+                            defaultCheckoutState,
+                            {
+                                ...defaultCheckoutActions,
+                                placeOrder: jest.fn(async () => Promise.reject(new Error(errorMessage)))
+                            }
+                        ),
+                        i18n,
+                        localVue,
+                        propsData,
+                        mocks: {
+                            $logger
+                        }
+                    });
+
+                    const toggleDialogErrorSpy = jest.spyOn(wrapper.vm, 'toggleDialogError');
+
+                    // A try-catch is needed because the exception otherwise is thrown and the test fails.
+                    try {
+                        // Act
+                        await wrapper.vm.submitOrder();
+                    } catch {
+                        // Assert
+                        expect(toggleDialogErrorSpy).toHaveBeenCalled();
+                    }
                 });
             });
         });