+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -4,6 +4,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.95.0
+------------------------------
+*April 26, 2021*
+
+### Added
+- Handling of duplicate orders.
+
+### Changed
+- Refactored handling order issues.
+
+
 v0.94.0
 ------------------------------
 *April 23, 2021*
@@ -26,6 +37,10 @@ v0.92.0
 
 ### Added
 - Use AuthToken phone nos. if checkout api fails to return a customer phone no.
+
+
+### Changed
+- Refactored handling order issues.
 
 
 v0.91.0